Menu menu functionality to its own file
